Updated makefile so it's more modern, fixed conversion issues for Windows, and added a workflow

+++ loadlib.c
@@ -168,10 +168,10 @@ static void setprogdir (lua_State *L) {
 
 
 static void pusherror (lua_State *L) {
-  int error = GetLastError();
+  int error = (int)GetLastError();
   char buffer[128];
   if (FormatMessageA(FORMAT_MESSAGE_IGNORE_INSERTS | FORMAT_MESSAGE_FROM_SYSTEM,
-      NULL, error, 0, buffer, sizeof(buffer)/sizeof(char), NULL))
+      NULL, (DWORD)error, 0, buffer, sizeof(buffer)/sizeof(char), NULL))
     lua_pushstring(L, buffer);
   else
     lua_pushfstring(L, "system error %d\n", error);
